Update router implementations.
