+++ src/sentry/taskworker/adapters.py
@@ -14,12 +14,15 @@ from typing import Generator
 from arroyo.backends.kafka import KafkaProducer
 from django.conf import settings
 from django.core.cache.backends.base import BaseCache
+from sentry_sdk import capture_exception
 from taskbroker_client.metrics import MetricsBackend, Tags
 from taskbroker_client.router import TaskRouter as LibraryRouter
 from taskbroker_client.types import AtMostOnceStore
 
+from sentry import options
 from sentry.conf.types.kafka_definition import Topic
-from sentry.taskworker.router import DefaultRouter
+from sentry.silo.base import SiloMode
+from sentry.utils import json
 from sentry.utils import metrics as sentry_metrics
 from sentry.utils.arroyo_producer import get_arroyo_producer
 from sentry.utils.memory import track_memory_usage as sentry_track_memory_usage
@@ -93,14 +96,30 @@ class SentryMetricsBackend(MetricsBackend):
             yield
 
 
-class SentryRouter(DefaultRouter, LibraryRouter):
-    """
-    Router that satisfies taskbroker_client's TaskRouter protocol while using
-    sentry's routing logic (settings, options, silo mode).
-    """
+class SentryRouter(LibraryRouter):
+    """Router that uses django settings and options to select topics at runtime."""
 
-    def route_namespace(self, name: str) -> str:  # type: ignore[override]
-        return super().route_namespace(name).value
+    def __init__(self) -> None:
+        routes = {}
+        if settings.TASKWORKER_ROUTES:
+            try:
+                routes = json.loads(settings.TASKWORKER_ROUTES)
+            except Exception as err:
+                capture_exception(err)
+        self._route_map = routes
+        self._default_topic = (
+            Topic.TASKWORKER_CONTROL
+            if SiloMode.get_current_mode() == SiloMode.CONTROL
+            else Topic.TASKWORKER
+        )
+
+    def route_namespace(self, name: str) -> str:
+        overrides = options.get("taskworker.route.overrides")
+        if name in overrides:
+            return Topic(overrides[name]).value
+        if name in self._route_map:
+            return Topic(self._route_map[name]).value
+        return self._default_topic.value
 
 
 _producer_local = threading.local()
